requirements.txt
APScheduler==3.9.1
deep-translator==1.8.3
requests==2.26.0
bs4==0.0.1
pandas==1.4.2
plotly==5.6.0
Pillow==9.0.0
Django==3.2
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-taggit==1.4.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
whitenoise==6.0.0
django-filter==21.1
django-import-export==2.8.0
django-adminlte3==0.1.6
django-bootstrap-v5==1.0.11

